refactor(schemas): retire settingsConfiguration re-export shim
The settingsConfiguration.ts shim was a pure pass-through re-export of the
canonical coreSettings surface, kept for legacy importers. No code imports
the shim path directly anymore, so inline its named exports into the
@shared/schemas barrel (preserving the exact public surface) and delete it.
